+++ test/support/test_backend.ex
@@ -1,6 +1,6 @@
 require Cldr.Currency.Backend
 
-defmodule Test.Cldr do
+defmodule MyApp.Cldr do
   use Cldr,
     locales: :all,
     default_locale: "en",